sangrahak-main/Major-/Backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
xgboost>=2.0.0
pymongo>=4.6.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
python-dotenv>=1.0.0
